moved file and fixed some include issues
